v3.2_Gemini/requirements.txt
# requirements.txt
faster-whisper
moviepy==1.0.3
ffmpeg-python
rich
openai
streamlit
python-dotenv